+++ src/js/main.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var load = require('./load');
-var pictures = require('./pictures');
+var drawPictures = require('./pictures');
 var upload = require('./upload');
 
-load('http://localhost:1507/api/pictures', pictures, '__jsonpCallback');
+load('http://localhost:1507/api/pictures', drawPictures, '__jsonpCallback');
